use same module structure as twitter gem
